refactor: rename panels to sidebar
