+++ src/webview/sidebar/inspector/index.ts
