+++ src/webview/sidebar/struct/structPinsModel.ts
